+++ agents-manage-ui/src/components/sidebar-nav/app-sidebar.tsx
@@ -15,7 +15,7 @@ import {
 } from 'lucide-react';
 import Link from 'next/link';
 import { useParams } from 'next/navigation';
-import type * as React from 'react';
+import { type ComponentProps, type Dispatch, type FC, useCallback } from 'react';
 import { MCPIcon } from '@/components/icons/mcp-icon';
 import { NavGroup } from '@/components/sidebar-nav/nav-group';
 import { ProjectSwitcher } from '@/components/sidebar-nav/project-switcher';
@@ -25,12 +25,11 @@ import {
   SidebarFooter,
   SidebarHeader,
   SidebarMenuButton,
-  SidebarRail,
-  useSidebar,
 } from '@/components/ui/sidebar';
 import { DOCS_BASE_URL } from '@/constants/page-descriptions';
 import { InkeepLogo } from '@/icons';
 import { cn } from '@/lib/utils';
+import { throttle } from '@/lib/utils/throttle';
 import type { NavItemProps } from './nav-item';
 
 const bottomNavItems: NavItemProps[] = [
@@ -47,9 +46,13 @@ const bottomNavItems: NavItemProps[] = [
   },
 ];
 
-export function AppSidebar(props: React.ComponentProps<typeof Sidebar>) {
+interface AppSidebarProps extends ComponentProps<typeof Sidebar> {
+  open: boolean;
+  setOpen: Dispatch<boolean>;
+}
+
+export const AppSidebar: FC<AppSidebarProps> = ({ open, setOpen, ...props }) => {
   const { tenantId, projectId } = useParams<{ tenantId: string; projectId?: string }>();
-  const { open } = useSidebar();
 
   const topNavItems: NavItemProps[] = projectId
     ? [
@@ -107,8 +110,21 @@ export function AppSidebar(props: React.ComponentProps<typeof Sidebar>) {
         },
       ];
 
+  const handleHover: NonNullable<ComponentProps<'div'>['onMouseEnter']> = useCallback(
+    throttle(200, (event) => {
+      setOpen(event.type === 'mouseenter');
+    }),
+    []
+  );
+
   return (
-    <Sidebar collapsible="icon" variant="inset" {...props}>
+    <Sidebar
+      collapsible="icon"
+      variant="inset"
+      onMouseEnter={handleHover}
+      onMouseLeave={handleHover}
+      {...props}
+    >
       <SidebarHeader>
         <SidebarMenuButton asChild>
           <Link href={`/${tenantId}/projects`}>
@@ -132,7 +148,6 @@ export function AppSidebar(props: React.ComponentProps<typeof Sidebar>) {
           <ProjectSwitcher />
         </SidebarFooter>
       )}
-      <SidebarRail />
     </Sidebar>
   );
-}
+};